requirements.txt
nltk
transformers
python-docx
langdetect
accelerate 
sentencepiece
dateparser